+++ internal/command/root_select_test.go
@@ -401,6 +401,22 @@ func TestRootCmd_Select_JSON(t *testing.T) {
 	t.Run("EmptyArray", selectTest(`[]`, "json", `.`, newline("[]"), nil))
 	t.Run("BlankInput", selectTest(``, "json", `.`, newline("{}"), nil))
 
+	t.Run("LengthSelector", selectTest(jsonData, "json", `.details.addresses.[#]`, newline("2"), nil))
+	t.Run("LengthDynamicSelector", selectTest(`{
+  "a": {
+    "id": 1,
+    "uses": [1]
+  },
+  "b": {
+    "id": 2,
+    "uses": [1, 2]
+  },
+  "c": {
+    "id": 3,
+    "uses": [1, 2, 3]
+  }
+}`, "json", `.(.uses.[#]=2).id`, newline("2"), nil))
+
 }
 
 func TestRootCmd_Select_YAML(t *testing.T) {